fix(ci): resolve 4 CI failures on PR #214
1. clippy: fix doc comment missing backticks (PostgreSQL, Neo4j, Qdrant);
   remove unused TenantId import from tombstoner test module

2. file size: split tenants/mod.rs (807 lines) into mod.rs + delete.rs —
   both now ≤600 lines per REQ-MD-01; delete_tenant handler moved to
   routes/tenants/delete.rs, re-exported via mod.rs

3. openapi sync: regenerate openapi.json from fresh print-openapi run;
   check-openapi-sync.sh passes

4. integration tests: add tombstone_producer: None to all AppState
   constructors in 5 integration test files (new field added in prior commit)

All checks: cargo clippy clean, 187 unit tests pass, file-size PASSED,
openapi-sync PASSED.
