+++ openapi.json
@@ -956,7 +956,7 @@
           "tenants"
         ],
         "summary": "Delete a tenant (owner-only, requires typed confirmation).",
-        "description": "Soft-deletes the tenant by setting `deleted_at` and transitioning its\nstatus to `deleting`, cancels all in-flight ingestion runs, then emits a\n`Tombstone` to `rb.tombstones.v1`. The tombstoner service performs the\nasync data-plane cleanup (PostgreSQL schema drop, Neo4j node removal,\nQdrant point deletion).\n\n**Idempotent**: repeating the call on a tenant already in `deleting` or\n`deleted` state returns `204 No Content`.\n\n**Typed confirmation**: the `X-Confirm` header must match the tenant slug\nexactly (case-insensitive) to prevent accidental deletions.\n\nReturns `503` if the Kafka producer is not available — the request can be\nretried once the broker is reachable.",
+        "description": "Soft-deletes the tenant by setting `deleted_at` and transitioning its\nstatus to `deleting`, cancels all in-flight ingestion runs, then emits a\n`Tombstone` to `rb.tombstones.v1`. The tombstoner service performs the\nasync data-plane cleanup (`PostgreSQL` schema drop, `Neo4j` node removal,\n`Qdrant` point deletion).\n\n**Idempotent**: repeating the call on a tenant already in `deleting` or\n`deleted` state returns `204 No Content`.\n\n**Typed confirmation**: the `X-Confirm` header must match the tenant slug\nexactly (case-insensitive) to prevent accidental deletions.\n\nReturns `503` if the Kafka producer is not available — the request can be\nretried once the broker is reachable.",
         "operationId": "delete_tenant",
         "parameters": [
           {
