+++ services/control-api/src/openapi.rs
@@ -5,7 +5,7 @@
 
 use utoipa::OpenApi;
 
-use crate::routes::{api_keys, audit, auth, auth_logout, auth_verify, github, health, ingest, me, query, repos, tenants};
+use crate::routes::{api_keys, audit, auth, auth_logout, auth_verify, github, health, ingest, me, query, repos, tenants, tenants::delete as tenant_delete};
 
 #[derive(OpenApi)]
 #[openapi(
@@ -33,7 +33,7 @@ use crate::routes::{api_keys, audit, auth, auth_logout, auth_verify, github, hea
         tenants::update_member_role,
         tenants::remove_member,
         tenants::transfer_ownership,
-        tenants::delete_tenant,
+        tenant_delete::delete_tenant,
         repos::connect_repo,
         repos::list_repos,
         repos::trigger_ingest,
@@ -74,7 +74,7 @@ use crate::routes::{api_keys, audit, auth, auth_logout, auth_verify, github, hea
             tenants::UpdateRoleRequest,
             tenants::UpdateRoleResponse,
             tenants::TransferOwnershipRequest,
-            tenants::DeleteTenantResponse,
+            tenant_delete::DeleteTenantResponse,
             repos::ConnectRepoRequest,
             repos::ConnectRepoResponse,
             repos::RepoItem,
