+++ services/control-api/tests/integration_tests.rs
@@ -39,6 +39,7 @@ fn test_state() -> AppState {
         gh: None,
         sse_bus: Arc::new(EventBus::new(SseConfig::default())),
         ingest_producer: None,
+        tombstone_producer: None,
         module_tree_cache: rb_query::new_module_tree_cache(),
     }
 }
@@ -331,6 +332,7 @@ async fn real_db_state() -> Option<(AppState, PgPool)> {
         gh: None,
         sse_bus: Arc::new(EventBus::new(SseConfig::default())),
         ingest_producer: None,
+        tombstone_producer: None,
         module_tree_cache: rb_query::new_module_tree_cache(),
     };
     Some((state, pool))
